feat: support multiple instruments at a time, change zoom factor
