+++ src/renderer/Pages/home.tsx
@@ -1,24 +1,24 @@
-import React, { useEffect, useRef, useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import fs from 'fs';
+import path from 'path';
 import { remote } from 'electron';
 import { useHistory } from 'react-router-dom';
-import path from 'path';
 import { useProject } from '../hooks/ProjectContext';
-import { LocalShim } from '../shims/LocalShim';
-import { PanelCanvas, PanelCanvasElement } from './PanelCanvas';
+import { PanelCanvas } from './PanelCanvas';
+import { Instrument, InstrumentFile, InstrumentFrame } from './Canvas/InstrumentFrame';
 
 const getInstruments = (instDirStr: string | undefined) => {
     if (instDirStr) {
         const instDir = fs.readdirSync(instDirStr);
 
-        const instrumentsTemp: instrumentType[] = [];
+        const instrumentsTemp: Instrument[] = [];
 
         instDir.forEach((file) => {
             const filePath = path.join(instDirStr, file);
             const fileStats = fs.statSync(filePath);
             if (fileStats.isDirectory()) {
                 const individualInstDir = fs.readdirSync(filePath);
-                const instrumentFiles: instrumentFileType[] = [];
+                const instrumentFiles: InstrumentFile[] = [];
                 individualInstDir.forEach((instrumentFile) => {
                     const instrumentFilePath = path.join(filePath, instrumentFile);
                     const instrumentFileStats = fs.statSync(instrumentFilePath);
@@ -43,98 +43,22 @@ const getInstruments = (instDirStr: string | undefined) => {
     return [];
 };
 
-type instrumentFileType = {
-    name: string,
-    path: string,
-    contents: string,
-}
-
-type instrumentType = {
-    name: string,
-    files: instrumentFileType[],
-}
-
 export const Home = () => {
     const { project, loadProject } = useProject();
-    const [instruments, setInstruments] = useState<instrumentType[]>([]);
-    const [selectedInstrument, setSelectedInstrument] = useState<instrumentType>({
-        name: '',
-        files: [],
-    });
+
+    const [availableInstruments, setAvailableInstruments] = useState<Instrument[]>([]);
+    const [selectedInstruments, setSelectedInstruments] = useState<Instrument[]>([]);
+
     const history = useHistory();
-    // useState(() => {
-    //     fs.watch(instDirStr, { encoding: 'utf8' }, (eventType, filename) => {
-    //         if (filename) {
-    //             console.log(filename);
-    //         }
-    //     });
-    // });
 
     useEffect(() => {
         if (project) {
             const bundlesPath = path.join(project.paths.project, project.paths.bundlesSrc);
-            setInstruments(getInstruments(bundlesPath));
+            setAvailableInstruments(getInstruments(bundlesPath));
         } else {
-            setInstruments(getInstruments(undefined));
+            setAvailableInstruments(getInstruments(undefined));
         }
     }, [project]);
-
-    const iframeRef = useRef<HTMLIFrameElement>();
-    const lastUpdate = Date.now();
-
-    useEffect(() => {
-        if (iframeRef.current && selectedInstrument.name) {
-            const iframeWindow = iframeRef.current.contentWindow;
-            const iframeDocument = iframeRef.current.contentDocument;
-
-            iframeDocument.body.style.overflow = 'hidden';
-
-            Object.assign(iframeRef.current.contentWindow, new LocalShim());
-
-            const rootTag = iframeDocument.createElement('div');
-            rootTag.id = 'ROOT_ELEMENT';
-
-            const mountTag = iframeDocument.createElement('div');
-            mountTag.id = 'MSFS_REACT_MOUNT';
-
-            rootTag.append(mountTag);
-
-            const pfdTag = iframeDocument.createElement('a35-x-ecam');
-            pfdTag.setAttribute('url', 'a?Index=1');
-
-            const scriptTag = iframeDocument.createElement('script');
-            scriptTag.text = selectedInstrument.files[1].contents;
-
-            const styleTag = iframeDocument.createElement('style');
-            styleTag.textContent = selectedInstrument.files[0].contents;
-
-            // Clear all intervals in the iframe
-            const lastInterval = iframeWindow.setInterval(() => {
-            }, 99999999);
-            for (let i = 0; i < lastInterval; i++) {
-                iframeWindow.clearInterval(i);
-            }
-
-            const lastTimeout = iframeWindow.setTimeout(() => {
-            }, 99999999);
-            for (let i = 0; i < lastTimeout; i++) {
-                iframeWindow.clearTimeout(i);
-            }
-
-            iframeDocument.head.innerHTML = '';
-            iframeDocument.body.innerHTML = '';
-            iframeDocument.body.style.margin = '0';
-
-            iframeDocument.body.append(rootTag);
-            iframeDocument.body.append(pfdTag);
-            iframeDocument.head.append(styleTag);
-            iframeDocument.head.append(scriptTag);
-
-            setInterval(() => {
-                iframeDocument.getElementById('ROOT_ELEMENT').dispatchEvent(new CustomEvent('update', { detail: Date.now() - lastUpdate }));
-            }, 50);
-        }
-    }, [iframeRef, selectedInstrument?.name, selectedInstrument.files, setSelectedInstrument]);
 
     return (
         <div className="w-full h-full flex">
@@ -170,10 +94,10 @@ export const Home = () => {
                     </button>
                 </div>
                 <div className="mt-4">
-                    {instruments.map((instrument) => (
+                    {availableInstruments.map((instrument) => (
                         <button
                             type="button"
-                            onClick={() => setSelectedInstrument(instrument)}
+                            onClick={() => setSelectedInstruments((insts) => [...insts, instrument])}
                         >
                             {instrument.name}
                         </button>
@@ -181,16 +105,11 @@ export const Home = () => {
                 </div>
             </div>
             <PanelCanvas render={(zoom) => (
-                selectedInstrument.name && (
-                    <PanelCanvasElement title={selectedInstrument.name} canvasZoom={zoom}>
-                        <iframe
-                            title="Instrument Frame"
-                            ref={iframeRef}
-                            width={768}
-                            height={1024}
-                        />
-                    </PanelCanvasElement>
-                )
+                <>
+                    {selectedInstruments.map((instrument) => (
+                        <InstrumentFrame selectedInstrument={instrument} zoom={zoom} />
+                    ))}
+                </>
             )}
             />
         </div>
